java documentation

+++ src/com/prog3/util/Float2.java
@@ -2,17 +2,21 @@ package com.prog3.util;
 
 import java.math.BigDecimal;
 
+/**
+ * Class for float operations.
+ */
 public class Float2{
   /**
-   * Round big decimal.
+   * Round float to decimalPlace
+   * es: 2.03124515 -> 2.03 if decimalPlace is 2.
    *
-   * @param d            the d
+   * @param d            the float number
    * @param decimalPlace the decimal place
-   * @return the big decimal
+   * @return rounded float
    */
-  public static BigDecimal round(float d, int decimalPlace) {
+  public static float round(float d, int decimalPlace) {
     BigDecimal bd = new BigDecimal(Float.toString(d));
     bd = bd.setScale(decimalPlace, BigDecimal.ROUND_HALF_UP);
-    return bd;
+    return bd.floatValue();
   }
 }
